backend/requirements.txt
anyio==3.4.0
asgiref==3.4.1
certifi==2021.10.8
cffi @ file:///C:/ci/cffi_1636542104204/work
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.70.1
h11==0.12.0
httptools==0.2.0
idna==3.3
importlib-metadata==4.10.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186163140/work
mkl-service==2.4.0
numpy==1.21.5
olefile==0.46
opencv-python==4.5.4.60
orjson==3.6.5
Pillow==8.4.0
pycparser
pydantic==1.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.26.0
six
sniffio==1.2.0
starlette==0.16.0
torch-cpu==1.10.0
torchvision-cpu==0.11.1
typing_extensions==4.0.1
ujson==4.3.0
urllib3==1.26.7
uvicorn==0.15.0
watchgod==0.7
websockets==10.1
wincertstore==0.2
zipp==3.6.0
